+++ CHANGELOG.md
@@ -188,4 +188,6 @@
 
 - Remove comments about coreui, edit readme (INDIGO Sprint 211015, [!188](https://github.com/TeskaLabs/asab-webui/pull/188))
 
+- Throw error for undefined response in MicroservicesListContainer (INDIGO Sprint 211029, [!189](https://github.com/TeskaLabs/asab-webui/pull/189))
+
 - Fix mock userinfo in dev locale (INDIGO Sprint 211029, [!191](https://github.com/TeskaLabs/asab-webui/pull/191))